Remove remnant verto screensharing and use kurento instead.
